IOS-1943 Moved saltpay's flatmap
